fix bug when len(ra) = 1 and len(radius) = 1
It appears that somewhere before here we coerce unit length radius
arrays to the length of the ra array. So, the only time we hit the code
I removed here is when the ra array is len 1. This segfaults because pt
is null.

N.B. I don't know this code particularly well so this almost certainly
isn't the best or most complete solution :)